refactor: provider user repository with db connection
