+++ internal/api/scim/server_test.go
@@ -23,7 +23,7 @@ var fixtures embed.FS
 func newServerFor(externalURL string) *Server {
 	return NewServer(&conf.GlobalConfiguration{
 		API: conf.APIConfiguration{ExternalURL: externalURL},
-	}, nil, NewUserMemoryRepository())
+	}, nil, NewUserRepository(nil))
 }
 
 func testFixture(t *testing.T, file string) string {
