+++ internal/api/scim_users_test.go
@@ -67,7 +67,7 @@ type SCIMUsersTestSuite struct {
 }
 
 func TestSCIMUsers(t *testing.T) {
-	users := scim.NewUserMemoryRepository()
+	users := scim.NewUserRepository()
 	api, config, err := setupAPIForTestWithCallback(func(cfg *conf.GlobalConfiguration, _ *storage.Connection) {
 		if cfg != nil {
 			cfg.Experimental.ScimEnabled = true
